isochronous tryings

+++ main.c
@@ -30,8 +30,8 @@ void parse_data(void* buffer, uint32_t length)
 	static uint64_t summ;
 	static uint64_t counter;
 	static struct timeval tv,tv_old,tv2,tv2_old;
-        uint64_t delta;
-        uint64_t speed=0;
+        double delta;
+        double speed=0.0;
 	summ+=length;
 	counter++;
 	if((tv2_old.tv_sec==0)&&(tv2_old.tv_usec==0)) gettimeofday(&tv2_old,NULL);
@@ -45,13 +45,13 @@ void parse_data(void* buffer, uint32_t length)
 		tv_old.tv_usec=tv.tv_usec;
 		gettimeofday(&tv,NULL);
 		delta=((tv.tv_sec-tv_old.tv_sec)*1000000)+(tv.tv_usec-tv_old.tv_usec);
-		speed=(uint64_t)((length*counter*1000000)/(delta));
+		speed=((length*counter*1000000)/(delta));
 	    
 	    counter=0;
 	    printf(" \033[0G\033[32;1m");
 	    hr_print(summ);
 	    printf("B \033[30;0mtransfered @ \033[35;1m");
-	    hr_print(speed);
+	    hr_print((uint64_t)speed);
 	    printf("B/s\033[30;0m  ");
 	    fflush(stdout);
 	}
